+++ napi/parser/raw-transfer/common.js
